requirements.txt
hydra-core==0.11.3
ipykernel==5.1.4
joblib==0.14.1
jupyter==1.0.0
matplotlib==3.1.3
networkx==2.4
numpy==1.18.1
pandas==1.0.1
pygraphviz==1.5
scikit-learn==0.22.1
scipy==1.4.1
tabulate==0.8.6
tensorboard==2.1.0
torch==1.4.0
torchvision==0.5.0
